# swagger/portal-swagger.yaml
swagger: '2.0'
info:
  version: '0.9'
  title: Portal
  description: >-
    The API of the Portal (the UI). All of these API calls may return different results
    depending on the current session; you will need to login to get access to restricted
    resources, e.g. to APIs or Content. Most of these end points require an
    `accept: application/json` header to return machine readable results, otherwise you
    will be served the corresponding HTML page. The data you get back from these end points
    corresponds 1:1 with the data visible on the pages; in most cases it's just that the
    rendering has been left out.
...
paths:
...
  /apis/{apiId}:
    get:
      description: >-
        Retrieve information on a specific API, including applications and subscriptions
      tags:
        - api
      parameters:
        - name: apiId
          required: true
          in: path
          description: The API ID to retrieve information on
          type: string
      responses:
        200:
          description: Successful response
...
  /admin/applications:
    get:
      description: Retrive a list of all applications (admin only)
      tags:
        - admin
      responses:
        200:
          description: Successful response
        403:
          description: Not allowed (only Admin)
...
  /applications/register:
    post:
      description: Register a new application
      tags:
        - application
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              appid:
                type: string
                description: ID of application to create
              appname:
                type: string
                description: Friendly name of application to create
                
      responses:
        201:
          description: Created
        400:
          description: Bad request
  /applications/{appId}:
    get:
      description: Retrieve information on an application; includes API credentials
      tags:
        - application
      parameters:
        - name: appId
          in: path
          required: true
          type: string
          description: The ID of the application to retrieve information on
      responses:
        200:
          description: Successful response
  /applications/{appId}/unregister:
    post:
      description: Unregisters (deletes) an application
      tags:
        - application
      parameters:
        - name: appId
          in: path
          required: true
          type: string
          description: The ID of the application to unregister (delete)
      responses:
        204:
          description: Successful response (no content)
        400:
          description: Bad request
        403:
          description: Not allowed
  /applications/{appId}/owners/add:
    post:
      description: Adds an owner (by email) to an application
      tags:
        - application
      parameters:
        - name: appId
          in: path
          required: true
          type: string
          description: The ID of the application to add an owner for
        - name: body
          in: body
          schema:
            type: object
            properties:
              owneremail:
                type: string
                description: Email address of owner to add (has to exist)
              ownerrole:
                type: string
                description: Role of owner
                enum:
                  - "owner"
                  - "collaborator"
                  - "reader"
      responses:
        201:
          description: Successfully added owner
  /applications/{appId}/owners/delete:
    post:
      description: Deletes an owner (by email) from an application
      tags:
        - application
      parameters:
        - name: appId
          in: path
          required: true
          type: string
          description: The ID of the application to delete an owner for
        - name: body
          in: body
          schema:
            type: object
            properties:
              owneremail:
                type: string
                description: Email address of owner to delete (has to exist)
      responses:
        200:
          description: Successfully deleted owner
  /application/{appId}/patch:
    post:
      description: Renames an application (only friendly name)
      tags:
        - application
      parameters:
        - name: appId
          in: path
          required: true
          type: string
          description: The ID of the application to patch the name of
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              appname:
                type: string
                description: New friendly name of application
      responses:
        200:
          description: Successfully patched name
  /application/{appId}/subscribe/{apiId}:
    get:
      description: Retrieve information on plans for a subscription
      tags:
        - application
      parameters:
        - name: appId
          in: path
          type: string
          required: true
          description: Application ID to add a subscription to
        - name: apiId
          in: path
          type: string
          required: true
          description: API ID to add a subscription to
      responses:
        200:
          description: Successful response
  
    post:
      description: Adds a subscription to an API for an Application
      tags:
        - application
      parameters:
        - name: appId
          in: path
          type: string
          required: true
          description: Application ID to add a subscription to
        - name: apiId
          in: path
          type: string
          required: true
          description: API ID to add a subscription to
        - name: body
          in: body
          required: true
          description: Subscription Parameters
          schema:
            type: object
            properties:
              application:
                type: string
                description: Application ID (once more)
              api:
                type: string
                description: API ID (once more)
              plan:
                type: string
                description: Plan ID to subscribe to (has to fit to API and user)
      responses:
        201:
          description: Successfully added subscription
  /application/{appId}/unsubscribe/{apiId}:
    post:
      description: Removes a subscription to an API for an Application
      tags:
        - application
      parameters:
        - name: appId
          in: path
          type: string
          required: true
          description: Application ID to add a subscription to
        - name: apiId
          in: path
          type: string
          required: true
          description: API ID to add a subscription to
      responses:
        204:
          description: Successfully unsubscribed to subscription
...
  /users/{userId}:
    get:
      description: Retrieve information on a user
      tags:
        - user
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user to retrieve information on
          type: string
      responses:
        200:
          description: Successful response
        403:
          description: Not allowed. You can only retrieve yourself if not admin
    post:
      description: Update a user, or delete password for a user
      tags:
        - user
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user to update
          type: string
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              __action:
                type: string
                description: >-
                  Set to `deletePassword` to delete a user's password, otherwise leave blank
              firstname:
                type: string
              lastname:
                type: string
              password:
                type: string
              __updategroups:
                type: boolean
                description: >-
                  Set to `true` to update the groups. Add the group names as properties in the body, containing the group name as value, e.g. `"dev":"dev"` or `"admin":"admin"`.
      responses:
        200:
          description: Successfully patched user.
        204:
          description: Successfully deleted password.
...